Remove unused code from network requester
