+++ service-providers/network-requester/Cargo.toml
@@ -10,7 +10,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bincode = "1.3"
 clap = "2.33.0"
 dirs = "3.0"
 futures = "0.3"
@@ -19,9 +18,8 @@ log = "0.4"
 pretty_env_logger = "0.4"
 publicsuffix = "1.5"
 reqwest = { version = "0.11", features = ["json"] }
-rocket = { version = "0.5.0-rc.1", features = ["json"], optional = true }
 serde = { version = "1.0", features = ["derive"] }
-sqlx = { version = "0.5", features = ["runtime-tokio-rustls", "sqlite", "macros", "migrate", "chrono"]}
+sqlx = { version = "0.5", features = ["chrono"]}
 thiserror = "1"
 tokio = { version = "1.19.1", features = [ "net", "rt-multi-thread", "macros", "time" ] }
 tokio-tungstenite = "0.14"
@@ -33,14 +31,8 @@ nymsphinx = { path = "../../common/nymsphinx" }
 ordered-buffer = {path = "../../common/socks5/ordered-buffer"}
 proxy-helpers = { path = "../../common/socks5/proxy-helpers" }
 socks5-requests = { path = "../../common/socks5/requests" }
+statistics = { path = "../../common/statistics" }
 websocket-requests = { path = "../../clients/native/websocket-requests" }
 
 [dev-dependencies]
 rand = "0.7"
-
-[build-dependencies]
-sqlx = { version = "0.5", features = ["runtime-tokio-rustls", "sqlite", "macros", "migrate"] }
-tokio = { version = "1.19.1", features = ["rt-multi-thread", "macros"] }
-
-[features]
-stats-service = ["rocket"]
