+++ service-providers/network-requester/src/statistics/comm.rs
@@ -4,7 +4,7 @@
 use futures::channel::mpsc;
 use futures::StreamExt;
 use log::*;
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use sqlx::types::chrono::{DateTime, Utc};
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
@@ -14,46 +14,12 @@ use tokio::sync::RwLock;
 use network_defaults::DEFAULT_NETWORK;
 use nymsphinx::addressing::clients::Recipient;
 use socks5_requests::{ConnectionId, RemoteAddress, Response};
+use statistics::{StatsMessage, StatsServiceData};
 
 use super::error::StatsError;
 
 const REMOTE_SOURCE_OF_STATS_PROVIDER_CONFIG: &str =
     "https://nymtech.net/.wellknown/network-requester/stats-provider.json";
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct StatsMessage {
-    pub stats_data: Vec<StatsServiceData>,
-    pub interval_seconds: u32,
-    pub timestamp: String,
-}
-
-impl StatsMessage {
-    pub fn to_bytes(&self) -> Result<Vec<u8>, StatsError> {
-        Ok(bincode::serialize(self)?)
-    }
-
-    #[cfg(feature = "stats-service")]
-    pub fn from_bytes(b: &[u8]) -> Result<Self, StatsError> {
-        Ok(bincode::deserialize(b)?)
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct StatsServiceData {
-    pub requested_service: String,
-    pub request_bytes: u32,
-    pub response_bytes: u32,
-}
-
-impl StatsServiceData {
-    pub fn new(requested_service: String, request_bytes: u32, response_bytes: u32) -> Self {
-        StatsServiceData {
-            requested_service,
-            request_bytes,
-            response_bytes,
-        }
-    }
-}
 
 #[derive(Clone, Debug)]
 pub struct StatsData {
