+++ service-providers/network-requester/src/statistics/mod.rs
@@ -5,5 +5,5 @@ mod comm;
 mod error;
 mod timer;
 
-pub use comm::{StatisticsCollector, StatisticsSender, StatsData, StatsMessage, StatsServiceData};
+pub use comm::{StatisticsCollector, StatisticsSender, StatsData};
 pub use timer::Timer;
